+++ sympl/_core/exceptions.py
@@ -1,3 +1,7 @@
+class UnitError(Exception):
+    pass
+
+
 class InvalidStateError(Exception):
     pass