+++ sympl/_core/units/_pint.py
@@ -16,10 +16,12 @@ class UnitRegistry(pint.UnitRegistry):
 unit_registry = UnitRegistry()
 unit_registry.define('degrees_north = degree_north = degree_N = degrees_N = degreeN = degreesN')
 unit_registry.define('degrees_east = degree_east = degree_E = degrees_E = degreeE = degreesE')
+unit_registry.define('degrees_Celsius = degC')
+unit_registry.define('degree_Celsius = degC')
 unit_registry.define('percent = 0.01*count = %')
 
 
-def units_are_compatible(unit1, unit2):
+def backend_units_are_compatible(unit1, unit2):
     """
     Determine whether a unit can be converted to another unit.
 
@@ -40,7 +42,7 @@ def units_are_compatible(unit1, unit2):
         return False
 
 
-def units_are_same(unit1, unit2):
+def backend_units_are_same(unit1, unit2):
     """
     Compare two unit strings for equality.
 
@@ -57,11 +59,11 @@ def units_are_same(unit1, unit2):
     return unit_registry(unit1) == unit_registry(unit2)
 
 
-def clean_units(unit_string):
+def backend_clean_units(unit_string):
     return str(unit_registry(unit_string).to_base_units().units)
 
 
-def is_valid_unit(unit_string):
+def backend_is_valid_unit(unit_string):
     """Returns True if the unit string is recognized, and False otherwise."""
     unit_string = unit_string.replace(
         '%', 'percent').replace(
@@ -74,17 +76,6 @@ def is_valid_unit(unit_string):
         return True
 
 
-def data_array_to_units(value, units):
-    if not hasattr(value, 'attrs') or 'units' not in value.attrs:
-        raise TypeError(
-            'Cannot retrieve units from type {}'.format(type(value)))
-    elif unit_registry(value.attrs['units']) != unit_registry(units):
-        attrs = value.attrs.copy()
-        value = unit_registry.Quantity(value, value.attrs['units']).to(units).magnitude
-        attrs['units'] = units
-        value.attrs = attrs
-    return value
+def backend_array_from_units_to_another(value, original_units, new_units):
+    return unit_registry.Quantity(value, original_units).to(new_units).magnitude
 
-
-def from_unit_to_another(value, original_units, new_units):
-    return (unit_registry(original_units)*value).to(new_units).magnitude